have decades appear in the container as cards and organize componenets
