+++ src/components/App/App.js
@@ -1,11 +1,13 @@
 import React, { Component } from "react";
-import Form from "./Form";
+import songData from "../../mockData"
+import Form from "../../Form";
+import DecadeContainer from "../DecadeContainer/DecadeContainer";
 
 class App extends Component {
   constructor() {
     super()
     this.state = {
-      oneHits: []
+      oneHits: songData
     }
   }
   render() {
@@ -13,6 +15,7 @@ class App extends Component {
       <div>
         <h1>One Hit Wonderzz</h1>
         <Form />
+        <DecadeContainer songData={this.state.oneHits}/>
       </div>
     )
   }
