+++ src/components/DecadeCard/DecadeCard.css
@@ -1,0 +1,6 @@
+.decade-card {
+    background-color: orange;
+    height: 20%;
+    width: 20%;
+    margin: 10px;
+}
