+++ src/components/DecadeContainer/DecadeContainer.css
@@ -1,0 +1,4 @@
+.decade-container {
+    display: flex;
+    flex-wrap: wrap;
+}
